Refactoring pin code input view
